sql: remove GLOB from Tarantool
GLOB is a legacy extension for LIKE from SQLite. As we want our SQL to
be close to ANSI SQL & LIKE to depend on collations, we do not want to
support it. This patch totally removes it from Tarantool along with any
mentions of it.

Part of #3589
Part of #3572
Needed for #3251
Needed for #3334